VMM: VMCPU_FF_VTIMER_ACTIVATED needs to be part of the per vCPU high priority pre execution mask, github:gh-34
VMCPU_FF_VTIMER_ACTIVATED is used to indicate to EM that the NEM execution loop should be called after a call
to VMR3WaitHalted() because the vTimer has likely expired and requires servicing. The flag is set at the end
of vmR3HaltMethod1Halt().
emR3ForcedAction() should normally clear the bit after it got set but because VMCPU_FF_VTIMER_ACTIVATED
is not part of VMCPU_FF_HIGH_PRIORITY_PRE_MASK the code to clear the bit is never reached and sub sequent
calls to VMR3WaitHalted() will cause an immediate exit because VMCPU_FF_VTIMER_ACTIVATED is still set and
it is part of VMCPU_FF_EXTERNAL_HALTED_MASK.
The observable effect is a high CPU usage for otherwise idling VMs because the vCPU threads never actually go
to sleep.



git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110295 cfe28804-0f27-0410-a406-dd0f0b0b656f