ValidationKit/utils/TestExecServ/TestExecServiceTcp.cpp: Fix possible crash due to a NULL pointer dereference when processing a REBOOT command in client/server mode
There is a race where the service has connected to the remote end through the client connect part while the listening server is not up yet.
If the remote part issues a REBOOT command txsTcpNotifyReboot() will be called which destroys the server and sets g_pTcpServer to NULL and
when txsTcpServerConnectThread() finally gets to execute it will pass NULL to RTTcpServerListen2() causing a crash while verifying the
magic. This was seen randomly for some testcases (especially Windows).
This is fixed by not destroying but only shutting the server down. The final destruction takes place in txsTcpTerm() after ensuring the
listener thread has finished.


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110650 cfe28804-0f27-0410-a406-dd0f0b0b656f