ValidationKit/tests/installation/tdGuestOsUnattendedInst1.py: Workaround for TestExecservice not starting after a reboot when the guest additions were installed
The observation on the testboxes was that TestExecService wasn't started after a
reboot when the guest additions were installed during the unattended installation.
This breaks testing completely obviously.
It turned out that Windows synchronizes time using NTP during the install but the
time on the testbox was running quite far behind (up to 2h because the testboxes are
using UTC but the Windows NTP sets the local time being CEST which is UTC+2h).
When the guest additions are installed the timesync service will synchronize host and guest
time on the first boot which causes a large jump backwards for the guest and can be observed
in the VBox.log as:

    05:42:16.721671 timesync vgsvcTimeSyncWorker: Radical guest time change: -25 188 575 203 200ns (GuestNow=1 759 210 936 721 671 800 ns GuestLast=1 759 236 125 296 875 000 ns fSetTimeLastLoop=true)

This causes the task scheduler service in the Windows guest to forget about the TestExecService task
it is supposed to start when vboxuser is automatically logged in. It is assumed that the task description
is not loaded because the file creation time is in the future (someone would need to verify that).
To workaround that this patch disables setting the host time when the guest service starts and increases
the threshold to try gradually synchronizing the guest time to 12h instead of the default 20min.



git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110700 cfe28804-0f27-0410-a406-dd0f0b0b656f